Fiona's suggested updates
Hey all!
I've added some pages: home-page; and then a page for each of our own blog pieces; a page that could have a form to fill in (share your own story page), and a page that I thought could just have some links on it to sites we've been using in the course, and maybe some social links.  I also added a css sheet that we can start updating.  So far, I've designed the pages so they have an identical header and menu. I hope this is ok!